requirements.txt
flask==1.1.2
flask-cors==3.0.10
gunicorn==20.0.4
pillow==8.0.1
scikit-image==0.17.2
numpy==1.19.2
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1+cpu
torchvision==0.8.2+cpu